scraper/requirements.txt
scrapy
jikanpy-v4
fuzzywuzzy[speedup]
python-dotenv
pymongo
beautifulsoup4
meilisearch
tqdm
pandas